Add support for waiting for a command to finish.
The idea of this is that you can write a secondary command that will
wait (probably asynchronusly or blocking rather than by polling) for a change
in whatever is being monitored.

When that secondary command exits, the main command (the one whose output
you are interested in) is run and the output displayed as usual, and then
the secondary command is spawned again.

In short, the secondary command (process['wait_command']['cmd']) and
process['args']
are run in turn, with only the output of task['args'] being output on screen.

I anticipate this will be good for things that don't change very frequently.
If process['wait_command']['cmd'] regularly quits immediately, the system could
get bogged down spawning new processes.  If that is the case, then it could be
mitigated by adding another Waiter implementation that is triggered by output
on stdout of the waiter process.  But ultimately if it is really fast changing
one might want to extend the program to do custom Python code in-process,
using the internal Twisted reactor to catch the errors in-process, and using
API calls/sockets/DBUS calls/whatever to update the display.